requirements.txt
# ─── Core Deep Learning ───────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12                   # DINO ViT-Small pretrained weights

# ─── Riemannian Geometry ──────────────────────────────────────────────────────
geoopt>=0.5.0                  # Stiefel + SPD manifold operations in PyTorch
geomstats>=2.7.0               # Statistical computations on manifolds

# ─── Bayesian & Uncertainty ───────────────────────────────────────────────────
torch-uncertainty>=0.2.0       # ECE, reliability diagrams, calibration metrics
laplace-torch>=0.1.2           # Laplace approximation / FIM estimation utilities

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python>=4.9.0
albumentations>=1.4.0          # Advanced augmentation pipeline
Pillow>=10.0.0
scikit-image>=0.22.0

# ─── Data & ML Utilities ──────────────────────────────────────────────────────
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.2.0
einops>=0.7.0                  # Clean tensor reshape/rearrange ops

# ─── Explainability ───────────────────────────────────────────────────────────
grad-cam>=1.5.0                # GradCAM++ for ViT
shap>=0.45.0                   # SHAP values for transformers

# ─── Experiment Tracking ──────────────────────────────────────────────────────
wandb>=0.16.0
omegaconf>=2.3.0               # Config management (YAML + override)

# ─── API Backend ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9        # File upload support in FastAPI
pydantic>=2.6.0

# ─── Model Export & Serving ───────────────────────────────────────────────────
onnx>=1.16.0
onnxruntime>=1.17.0            # CPU-optimized inference
huggingface-hub>=0.22.0        # Model card + HF Spaces deployment

# ─── Federated Learning ───────────────────────────────────────────────────────
flwr>=1.8.0                    # Flower federated learning framework

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0            # Property-based testing

# ─── Code Quality ─────────────────────────────────────────────────────────────
black>=24.0.0
isort>=5.13.0
mypy>=1.9.0
ruff>=0.3.0

# ─── Visualization ────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0

# ─── Notebook Support ─────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# ─── Docs ─────────────────────────────────────────────────────────────────────
mkdocs>=1.5.0
mkdocs-material>=9.5.0
